Remember step, create component walkthrough could be move to rcl
