+++ package.json
@@ -7,7 +7,7 @@
     "@emotion/react": "^11.13.3",
     "@emotion/styled": "^11.13.0",
     "@mui/material": "^6.1.5",
-    "pankosmia-rcl": "^0.1.43",
+    "pankosmia-rcl": "0.1.45",
     "pithekos-lib": "^0.11.20",
     "react": "^17.0.0 || ^18.0.0",
     "react-dom": "^17.0.0 || ^18.0.0",
