+++ pankosmia_metadata.json
@@ -3,6 +3,7 @@
   "require": {
     "net": false
   },
+  "minServerVersion": "0.14.12",
   "i18n": {
     "title": {
       "ar": "الفهرس",
